Resolved bug related to Game init() and useAssistant() methods synchronization.
Added an attribute to constructor of ConnectionStatusChange.

Resolved bug related to deregistering connections from lobby, now it works fine. Each client gets notified of disconnection by lobbyChange.

Changed parameters type in Game init() from ArrayList to List to make it flexible.

Resolved bug in deregisterConnection() and addToLobby() in server.

Implemented method createGame() in server, but it is still to debug because it gives errors.