Added coin to player for each 3rd student of particular color in dining room.
Resolved bug related to client that tries to enter in game when there is no free space for him in game.

Corrected startTurn() of ClientController so that when it receives EndOfGameChange, it can exit from all recursive calls.

Added isGameOn variable to ClientGameBoard in order to manage game end.

Corrected ConnectionTracker implementation so that it executes until Connection gets deactivated

Corrected implementation of EndOfGameChange and GameBoardChange so that they set isGameOn variable of ClientGameBoard.