Started working on connection resilience but there are still bugs to fix. Added player activation and disactivation in VirtualView and synchronized it with Game methods so that client can (de)activate player only when Game thread is in wait() for other client's action
Added GameBoardChange execution when client tries to reconnect with the same name and server accepts it.

Changed a little checkEndGame() implementation of Game so that it sends EndOfGameChange to all active clients.

Changed compareTo implementation of Player in order to put not active players at the end of the array