Finished connection resilience. There still may be some bugs to resolve.
Resolved with reconnected client that had no VirtualView attached to its connection.

Reimplemented refillClouds() and useCloud() methods in Game so that they are more appropriate for connection resilience.

Fixed a bug in moveMotherNature() in Character, Character8 in order to not allow negative value

Inserted a check isGameOn in ClientController each time before doing an action

Changed name insertion logic so that only uppercase strings are used