Improved checkEndGame() method of Game.
Fixed bug in Character.buy() in order to leave 1 coin on the character card if a player buys it for the first time.

Synchronized buyCharacter() and activateCharacter() methods and inserted notifyAll() in activateCharacter().

Resolved bug related to refill of those characters that have students on card.

Added isGameOn() control and server exceptions catch inside ClientController.

Added parameters check and exception throw in Game and GameBoard.

Added print of teamColor in 4-player game.

Highlighted the player frame with red color of non active player in cli version of game.

Added model print when a player changes its connectionStatus