Resolved bug with connection resilience.
Now the server holds the packet sent from client while game is suspended and sends it only after game gets unsuspended.

On the client side, waitModelChange() in ClientConnection.send() and ClientController.startTurn() stamps the GameSuspended and GameReactivated exceptions and discards them until EndOfChangesException arrives.

Resolved bug in Character2 related to lack of control of the currentPlayer